+++ package.json
@@ -10,6 +10,8 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/stack": "^6.3.29",
     "react": "18.2.0",
     "react-native": "0.74.1",
     "react-native-material-ui": "^1.30.1",
